Final version for illiquid bid ask-spreads that are $0.05-$0.10
This is the final copy with correction for if the stock goes below 50 cents. Used to be an infinite loop but that was addressed. As far as I'm concerned, for the illiquid bid-asks this works well. The next direction is to price options locally and make decisions based on that. I'll also figure out a way to "Monte-Carlo" this and determine some kind of expected value although I am seeing a lot of variance